+++ public/assets/scripts/choices.search-basic.mjs
@@ -4508,7 +4508,7 @@ var Choices = /** @class */ (function () {
         var callbackOnCreateTemplates = this.config.callbackOnCreateTemplates;
         var userTemplates = {};
         if (typeof callbackOnCreateTemplates === 'function') {
-            userTemplates = callbackOnCreateTemplates.call(this, strToEl, escapeForTemplate);
+            userTemplates = callbackOnCreateTemplates.call(this, strToEl, escapeForTemplate, getClassNames);
         }
         var templating = {};
         Object.keys(this._templates).forEach(function (name) {