+++ src/scripts/interfaces/options.ts
@@ -3,6 +3,7 @@ import { InputChoice } from './input-choice';
 import { ClassNames } from './class-names';
 import { PositionOptionsType } from './position-options-type';
 import { Types } from './types';
+import { Templates } from './templates';
 
 export const ObjectsInConfig: string[] = ['fuseOptions', 'classNames'];
 
@@ -611,7 +612,7 @@ export interface Options {
    *
    * @default null
    */
-  callbackOnCreateTemplates: ((template: Types.StrToEl, escapeForTemplate: Types.EscapeForTemplateFn) => void) | null;
+  callbackOnCreateTemplates: ((template: Types.StrToEl, escapeForTemplate: Types.EscapeForTemplateFn, getClassNames: Types.GetClassNamesFn) => Partial<Templates>) | null;
 
   appendGroupInSearch: false;
 }