+++ src/scripts/interfaces/types.ts
@@ -4,6 +4,7 @@ import { StringPreEscaped } from './string-pre-escaped';
 export namespace Types {
   export type StrToEl = (str: string) => HTMLElement | HTMLInputElement | HTMLOptionElement;
   export type EscapeForTemplateFn = (allowHTML: boolean, s: StringUntrusted | StringPreEscaped | string) => string;
+  export type GetClassNamesFn = (s: string | Array<string>) => string;
   export type StringFunction = () => string;
   export type NoticeStringFunction = (value: string, valueRaw: string) => string;
   export type NoticeLimitFunction = (maxItemCount: number) => string;